Ana Răus: Modificare: adaugare max pentru unele coloane

+++ classes/instructor.py
@@ -6,14 +6,14 @@ from sqlalchemy import Column, String, Integer, Date, Table, ForeignKey
 class Instructor(Base):
     __tablename__ = 'instructor'
 
-    id = Column(Integer, primary_key=True)
+    id = Column(Integer(100), primary_key=True)
 
     #One to One toate
     personal = relationship("Personal", back_populates="instructor")
-    personal_id = Column(Integer, ForeignKey("personal.id"))
+    personal_id = Column(Integer(100), ForeignKey("personal.id"))
     vehicul = relationship("Vehicul", back_populates="instructor")
-    vehicul_id = Column(Integer, ForeignKey("vehicul.id"))
+    vehicul_id = Column(Integer(100), ForeignKey("vehicul.id"))
     cont = relationship("Cont", back_populates="instructor")
-    cont_id = Column(Integer, ForeignKey("cont.id"))
+    cont_id = Column(Integer(100), ForeignKey("cont.id"))
     programare = relationship("Programare")
